# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  username: string;
  displayName: string;
  dob: string;
  avatarColor: string;
  winRate: number;
  currentStreak: number;
  points: number;
  isBankrupt: boolean;
  previousWagers: number;
  /** @nullable */
  lastDailyClaim: Date | null;
  profileComplete: boolean;
}