# lib/api-zod/src/generated/types/wager.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Wager {
  id: string;
  userId: string;
  fixtureId: string;
  choice: string;
  question: string;
  prediction: string;
  amount: number;
  odds: number;
  potentialPayout: number;
  status: string;
  payout: number;
  createdAt: Date;
  /** @nullable */
  settledAt: Date | null;
}