# lib/api-zod/src/generated/types/wagerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WagerInput {
  /** @minLength 1 */
  fixtureId: string;
  /** @minLength 1 */
  choice: string;
  question?: string;
  prediction?: string;
  /** @minimum 1 */
  amount: number;
  /** @minimum 1 */
  odds: number;
}