# lib/api-zod/src/generated/types/wagerResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";
import type { Wager } from "./wager";

export interface WagerResult {
  user: User;
  wager: Wager;
}